Solved page changing problem.
Deleted entranceWindow and registerWindow widgets and their files
permanently. Created a widget containing a stackedWidget and copied the
pages inside of it. That way page switching will be solved. However, it
could possibly make things complicated as I add new pages. This is
probably not the right way to go considering Seperation of Concerns
principle.